--- src/runners/commander/types.ts
@@ -1,6 +1,0 @@
-export type CliOptions = {
-  template: string;
-  importAlias: string;
-  packageManager: string;
-  resetPreferences: boolean;
-};
